--- client-extensions/liferay-sample-custom-element-4/assets/style.css
@@ -1,5 +1,0 @@
-.greeting {
-	background-color: #f3e5ab;
-	display: block;
-	padding: 10px;
-}
