--- client-extensions/liferay-sample-custom-element-4/client-extension.dev.yaml
@@ -1,4 +1,0 @@
-liferay-sample-custom-element-4:
-    baseURL: http://localhost:3000
-    urls:
-        - index.js
