--- client-extensions/liferay-sample-custom-element-4/client-extension.yaml
@@ -1,13 +1,0 @@
-assemble:
-    - from: build/static
-      into: static
-liferay-sample-custom-element-4:
-    friendlyURLMapping: liferay-sample-custom-element-4
-    htmlElementName: liferay-sample-custom-element-4
-    instanceable: false
-    name: Liferay Sample Custom Element 4
-    portletCategoryName: category.client-extensions
-    type: customElement
-    urls:
-        - index.*.js
-    useESM: true
