--- client-extensions/liferay-sample-custom-element-4/package.json
@@ -1,15 +1,0 @@
-{
-	"devDependencies": {
-		"css-loader": "6.10.0",
-		"style-loader": "3.3.4",
-		"webpack": "5.90.1",
-		"webpack-cli": "5.1.4"
-	},
-	"name": "@liferay/liferay-sample-custom-element-4",
-	"private": true,
-	"scripts": {
-		"build": "webpack",
-		"start": "webpack-dev-server"
-	},
-	"version": "0.0.0"
-}
