+++ exercises/module-1/App.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import TicketsList from './assets/components/TicketsList';
+
+function App() {
+  return (
+    <div className="App">
+      <TicketsList />
+    </div>
+  );
+}
+
+export default App;
