+++ exercises/module-1/assets/components/TicketsList.js
@@ -1,6 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import ReactDOM from 'react-dom';
-import './style.css';
+import '../style.css';
+import Ticket from './Ticket';
+
+
 
 // API helper function
 const api = async (url, options = {}) => {
@@ -20,42 +23,7 @@ const getCurrentUserId = async () => {
 	return data.id;
 };
 
-// Ticket component
-const Ticket = React.memo(({ ticket, onToggle, isOpen }) => (
-	<div className="ticket-card" onClick={() => onToggle(ticket.id)}>
-		<div className="ticket-header">
-			<h3 className="ticket-subject">{ticket.subject}</h3>
-			<span className={`ticket-status ${ticket.priority?.key}`}>
-				{ticket.ticketStatus.name}
-			</span>
-		</div>
-		{isOpen && (
-			<div className="ticket-details">
-				<p>Description: {ticket.description}</p>
-				{ticket.attachment && (
-					<a
-						href={ticket.attachment.link.href}
-						target="_blank"
-						rel="noopener noreferrer"
-						className="ticket-attachment"
-					>
-						View Attachment
-					</a>
-				)}
-				{ticket.relatedTickets && ticket.relatedTickets.length > 0 && (
-					<div className="related-tickets">
-						<h4>Related Tickets</h4>
-						{ticket.relatedTickets.map((related) => (
-							<div className="related-ticket-item" key={related.id}>
-								ID: {related.id} - {related.subject}
-							</div>
-						))}
-					</div>
-				)}
-			</div>
-		)}
-	</div>
-));
+
 
 // Main component to display tabs and tickets list
 function TicketsList() {
@@ -135,43 +103,5 @@ function TicketsList() {
 	);
 }
 
-// Custom Element class
-class CustomElement extends HTMLElement {
-	connectedCallback() {
-		// Ensure the React component is rendered only once
-		if (!this.rendered) {
-			// Create a container if it doesn't exist
-			const container = document.createElement('div');
-			container.id = 'tickets-root';
-			this.appendChild(container);
 
-			// Render the React component into the container
-			ReactDOM.render(<TicketsList />, container);
-			this.rendered = true;
-		}
-	}
-
-	disconnectedCallback() {
-		// Clean up the React component when the element is removed
-		const container = this.querySelector('#tickets-root');
-		if (container) {
-			ReactDOM.unmountComponentAtNode(container);
-		}
-		this.rendered = false;
-	}
-}
-
-// Define the custom element
-const ELEMENT_NAME = 'liferay-clarity-custom-element';
-
-if (!customElements.get(ELEMENT_NAME)) {
-	customElements.define(ELEMENT_NAME, CustomElement);
-}
-
-// Automatically add the custom element to the page if not already present
-document.addEventListener('DOMContentLoaded', () => {
-	if (!document.querySelector(ELEMENT_NAME)) {
-		const customElement = document.createElement(ELEMENT_NAME);
-		document.body.appendChild(customElement);
-	}
-});
+export default TicketsList;
