+++ exercises/module-1/public/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Clarity Ticketing UI</title>
+  </head>
+  <body>
+    <root></root>
+    <tickets-root></tickets-root>
+    <script type="module" src="build/static/index.js"></script>
+  </body>
+</html>
